seperated into main and header
