+++ test_coords.csv
@@ -1,0 +1,6 @@
+﻿long, lat, time
+-37.673467 ,144.900807,100
+-37.675475, 145.244541,100
+-37.933589, 144.913295,100
+-37.931713, 145.237128,100
+-37.833589, 145.0,10
